make the registry pullable
To make the registry pullable by tags the droplet is saved with
guid and with the digest. That way kube is able to pull the images
when going through the docker manifest. Another way would be to pull
images by digest. This, however, would require to save the digest of
the manifest, as kube/docker is comparing the digest of the manifest
with the request digest. A todo would definitely be to avoid saving
droplet redundantly.